Semi-functional route importer and unhashing, but disabled drawing temporarily

+++ FoxKit/Assets/Fox/Core/DynamicProperties/DynamicProperty_DynamicArray_Path.cs
@@ -12,8 +12,8 @@ namespace Fox.Core
         internal override PropertyInfo GetPropertyInfo() => new PropertyInfo(Name, PropertyInfo.PropertyType.Path, 0, container: GetContainerType());
         
         public override Value GetValue() => new Value(SerializedField);
-        public override Value GetElement(ushort index) => new Value(SerializedField[index]);
+        public override Value GetElement(int index) => new Value(SerializedField[index]);
 
-        public override void SetElement(ushort index, Value value) => SerializedField.Insert(index, value.GetValueAsPath());
+        public override void SetElement(int index, Value value) => SerializedField.Insert(index, value.GetValueAsPath());
     }
 }
