+++ FoxKit/Assets/Fox/Core/DynamicProperties/DynamicProperty_StaticArray_uint64.cs
@@ -22,9 +22,9 @@ namespace Fox.Core
         }
 
         public override Value GetValue() => new Value(SerializedField);
-        public override Value GetElement(ushort index) => new Value(SerializedField[index]);
+        public override Value GetElement(int index) => new Value(SerializedField[index]);
 
-        public override void SetElement(ushort index, Value value)
+        public override void SetElement(int index, Value value)
         {
             SerializedField[index] = value.GetValueAsUInt64();
         }
